chore(auth): logical separation
