+++ packages/auth/__tests__/client/utils/passkey.test.ts
@@ -1,45 +1,15 @@
 import {
-	convertArrayBufferToBase64Url,
-	convertBase64UrlToArrayBuffer,
-} from '../../src/utils/passkey/base64Url';
-import {
 	deserializeJsonToPkcCreationOptions,
 	serializePkcToJson,
-} from '../../src/utils/passkey/serde';
+} from '../../../src/client/utils/passkey/serde';
 import {
 	passkeyRegistrationRequest,
 	passkeyRegistrationRequestJson,
 	passkeyRegistrationResult,
 	passkeyRegistrationResultJson,
-} from '../mockData';
+} from '../../mockData';
 
 describe('passkey', () => {
-	it('converts ArrayBuffer values to base64url', () => {
-		expect(convertArrayBufferToBase64Url(new Uint8Array([]))).toBe('');
-		expect(convertArrayBufferToBase64Url(new Uint8Array([0]))).toBe('AA');
-		expect(convertArrayBufferToBase64Url(new Uint8Array([1, 2, 3]))).toBe(
-			'AQID',
-		);
-	});
-	it('converts base64url values to ArrayBuffer', () => {
-		expect(
-			convertArrayBufferToBase64Url(convertBase64UrlToArrayBuffer('')),
-		).toBe(convertArrayBufferToBase64Url(new Uint8Array([])));
-		expect(
-			convertArrayBufferToBase64Url(convertBase64UrlToArrayBuffer('AA')),
-		).toBe(convertArrayBufferToBase64Url(new Uint8Array([0])));
-		expect(
-			convertArrayBufferToBase64Url(convertBase64UrlToArrayBuffer('AQID')),
-		).toBe(convertArrayBufferToBase64Url(new Uint8Array([1, 2, 3])));
-	});
-
-	it('converts base64url to ArrayBuffer and back without data loss', () => {
-		const input = '_h7NMedx8qUAz_yHKhgHt74P2UrTU_qcB4_ToULz12M';
-		expect(
-			convertArrayBufferToBase64Url(convertBase64UrlToArrayBuffer(input)),
-		).toBe(input);
-	});
-
 	it('serializes pkc into correct json format', () => {
 		expect(JSON.stringify(serializePkcToJson(passkeyRegistrationResult))).toBe(
 			JSON.stringify(passkeyRegistrationResultJson),
