+++ packages/auth/src/client/apis/index.ts
@@ -1,0 +1,1 @@
+export { associateWebAuthnCredential } from './associateWebAuthnCredential';
