+++ packages/auth/src/client/types/index.ts
@@ -1,0 +1,1 @@
+export { AssociateWebAuthnCredentialOutput } from './outputs';
