+++ packages/auth/src/client/types/outputs.ts
@@ -1,0 +1,6 @@
+/**
+ * Output type for Cognito associateWebAuthnCredential API.
+ */
+export interface AssociateWebAuthnCredentialOutput {
+	credentialId?: string;
+}
