+++ packages/auth/src/client/utils/passkey/errors.ts
