+++ packages/auth/src/client/utils/passkey/getIsPasskeySupported.native.ts
