+++ packages/auth/src/client/utils/passkey/getIsPasskeySupported.ts
