+++ packages/auth/src/foundation/convert/index.ts
@@ -1,0 +1,4 @@
+export {
+	convertArrayBufferToBase64Url,
+	convertBase64UrlToArrayBuffer,
+} from './base64url';
