+++ packages/auth/src/providers/cognito/index.ts
@@ -27,7 +27,6 @@ export { rememberDevice } from './apis/rememberDevice';
 export { forgetDevice } from './apis/forgetDevice';
 export { fetchDevices } from './apis/fetchDevices';
 export { autoSignIn } from './apis/autoSignIn';
-export { associateWebAuthnCredential } from './apis/associateWebAuthnCredential';
 export {
 	ConfirmResetPasswordInput,
 	ConfirmSignInInput,
@@ -63,7 +62,6 @@ export {
 	UpdateUserAttributeOutput,
 	SendUserAttributeVerificationCodeOutput,
 	FetchDevicesOutput,
-	AssociateWebAuthnCredentialOutput,
 } from './types/outputs';
 export {
 	AuthUser,
