+++ packages/auth/src/utils/index.ts
@@ -4,4 +4,3 @@
 export { getAuthUserAgentDetails } from './getAuthUserAgentDetails';
 export { getAuthUserAgentValue } from './getAuthUserAgentValue';
 export { openAuthSession } from './openAuthSession';
-export { registerPasskey } from './passkey';
